Some cleanup and new indirect left recursion test
