+++ grammar_parser/tests/newGrammar_test_only_dont_modify.dsl
@@ -69,3 +69,6 @@
 
 <test_indirect_LR_num> = <test_indirect_LR_expr>;
 <test_indirect_LR_expr> = (<test_indirect_LR_num>, '-', <test_LR_num>) / <test_LR_num>; # Should match 0-0-0-0-0-0-0-0 etc #
+
+<test_term> = (<test_term>, "+", <test_fact>)/(<test_term>, "-", <test_fact>)/<test_fact>;
+<test_fact> = (<test_fact>, "*", <Num>)/(<test_fact>, "/", <Num>)/<Num>;
