+++ parser_core/src/var_name/_var_name_indirect.rs
@@ -8,6 +8,7 @@ use std::u32;
 
 // While complex still under development and the core of the entire program is here so complexity is
 // Acceptable
+use crate::publisher_utils::*;
 use crate::source::Source;
 use crate::Context;
 use cache::ASTOrLR;
@@ -18,214 +19,10 @@ use cache::LR;
 use publisher::Publisher;
 use rules::{Key, Rules};
 
-pub fn _var_name<T: Cache, S: Publisher>(
-    rule: Rules,
-    context: &Context<T, S>,
-    func: fn(&Context<T, S>, &Source, u32) -> (bool, u32, AST),
-) -> impl Fn(&Source, u32) -> (bool, u32, AST) + '_ {
-    move |source: &Source, position: u32| {
-        _var_name_kernel_direct_lr(rule, context, source, position, func)
-    }
-}
-
 // Memo(R, P) -> MemoEntry in paper
 // In our case Cache(R, P, ...) -> Option<(Option<bool>, u32, Key)> where Key is essentially the AST/ which in the paper can have
 // a special value of FAIL, I used the option in bool instead which is wrong.
 // Ideally I'd use Cache(R, P) -> Option<(bool, u32, Option<Key>)>
-
-fn publisher_setup_node<T: Cache, S: Publisher>(
-    context: &Context<T, S>,
-    rule: Rules,
-) -> (Option<Key>, Key) {
-    let mut publisher = context.stack.borrow_mut();
-    let parent_key: Option<Key> = publisher.last_node();
-    let current_key = publisher.add_node(rule, 0, 0, false);
-    publisher.set_last_node(Some(current_key));
-    return (parent_key, current_key);
-}
-
-fn publisher_get_last_node<T: Cache, S: Publisher>(context: &Context<T, S>) -> Option<Key> {
-    let publisher = context.stack.borrow_mut();
-    let parent_key: Option<Key> = publisher.last_node();
-    return parent_key;
-}
-
-fn publisher_update_node<T: Cache, S: Publisher>(
-    context: &Context<T, S>,
-    start_position: u32,
-    end_position: u32,
-    is_true: bool,
-    parent_key: Option<Key>,
-    current_key: Key,
-) {
-    let mut publisher = context.stack.borrow_mut();
-    match parent_key {
-        None => {}
-        Some(pkey) => {
-            // REMOVE THIS IF STATEMENT IF YOU ALSO WANT TO SEE ERRORS,
-            // E.G FOR LINTERS SO THEY CAN FIND THE LONGEST MATCH AND SUGGEST A FIX.
-            if is_true {
-                publisher.connect(pkey, current_key);
-            }
-        }
-    }
-    publisher.set_node_start_position(current_key, start_position);
-    publisher.set_node_end_position(current_key, end_position);
-    publisher.set_node_result(current_key, is_true);
-    publisher.set_last_node(parent_key);
-}
-fn publisher_update_node_no_set_last_node<T: Cache, S: Publisher>(
-    context: &Context<T, S>,
-    start_position: u32,
-    end_position: u32,
-    is_true: bool,
-    parent_key: Option<Key>,
-    current_key: Key,
-) {
-    let mut publisher = context.stack.borrow_mut();
-    match parent_key {
-        None => {}
-        Some(pkey) => {
-            // REMOVE THIS IF STATEMENT IF YOU ALSO WANT TO SEE ERRORS,
-            // E.G FOR LINTERS SO THEY CAN FIND THE LONGEST MATCH AND SUGGEST A FIX.
-            if is_true {
-                publisher.connect(pkey, current_key);
-            }
-        }
-    }
-    publisher.set_node_start_position(current_key, start_position);
-    publisher.set_node_end_position(current_key, end_position);
-    publisher.set_node_result(current_key, is_true);
-}
-
-fn _var_name_kernel_basic<T: Cache, S: Publisher>(
-    rule: Rules,
-    context: &Context<T, S>,
-    source: &Source,
-    position: u32,
-    func: fn(&Context<T, S>, &Source, u32) -> (bool, u32, AST),
-) -> (bool, u32, AST) {
-    println!("{:?} Entering Var_Name Kernel", rule);
-
-    {
-        let mut cache: std::cell::RefMut<T> = context.cache.borrow_mut();
-        // Removes value from hashmap so we need to reinsert afterwards.
-        let memo_entry = cache.check_lr(rule, position);
-        match memo_entry {
-            Some(memo_entry) => {
-                // START Publisher: If cached we connect the parent to the existing key.
-                let current_key: AST = memo_entry.ast_or_lr.into();
-                let current_key: Key = current_key.into();
-                publisher_update_node(
-                    context,
-                    position,
-                    memo_entry.position,
-                    memo_entry.is_true,
-                    publisher_get_last_node(context),
-                    current_key,
-                );
-                // END PUBLISHER
-
-                // Cached Values do exist.
-                println!("{:?} Entering Cached Value Block", rule);
-                println!("{:?} Exiting Cached Value Block", rule);
-                println!("{:?} Exiting Var_Name Kernel", rule);
-                return (
-                    memo_entry.is_true,
-                    memo_entry.position,
-                    memo_entry.ast_or_lr.into(),
-                );
-            }
-            None => {}
-        }
-    }
-    // Publisher: If not cached we need to create a Node prior to calling the function, then connect parent to the returned key of func.
-    let (parent_key, current_key) = publisher_setup_node(context, rule);
-    // If not cached value we just run the func.
-    println!("{:?} Entering No Cached Value Block", rule);
-    let ans = func(context, source, position);
-    // Publisher, we then update the node which sets it's values and connects the parent to the current key.
-    publisher_update_node(context, position, ans.1, ans.0, parent_key, current_key);
-    let mut cache = context.cache.borrow_mut();
-    cache.push(
-        rule,
-        ans.0,
-        position,
-        ans.1,
-        cache::ASTOrLR::AST(AST::SUCCESS(current_key)),
-    );
-    println!("{:?} Exiting No Cached Value Block", rule);
-    println!("{:?} Exiting Var_Name Kernel", rule);
-    return (ans.0, ans.1, AST::SUCCESS(current_key));
-}
-
-fn _var_name_kernel_deny_lr<T: Cache, S: Publisher>(
-    rule: Rules,
-    context: &Context<T, S>,
-    source: &Source,
-    position: u32,
-    func: fn(&Context<T, S>, &Source, u32) -> (bool, u32, AST),
-) -> (bool, u32, AST) {
-    println!("{:?} Entering Var_Name Kernel", rule);
-
-    {
-        let mut cache: std::cell::RefMut<T> = context.cache.borrow_mut();
-        // Removes value from hashmap so we need to reinsert afterwards.
-        let memo_entry = cache.check_lr(rule, position);
-        match memo_entry {
-            Some(memo_entry) => {
-                // Cached Values do exist.
-                println!("{:?} Entering Cached Value Block", rule);
-                println!("{:?} Exiting Cached Value Block", rule);
-                println!("{:?} Exiting Var_Name Kernel", rule);
-                if memo_entry.ast_or_lr == ASTOrLR::AST(AST::FAIL) {
-                    println!("{:?} Left Recursion was detected! Aborting", rule);
-                    exit(-1);
-                }
-                // START Publisher: If cached we connect the parent to the existing key.
-                let current_key: AST = memo_entry.ast_or_lr.into();
-                let current_key: Key = current_key.into();
-                publisher_update_node(
-                    context,
-                    position,
-                    memo_entry.position,
-                    memo_entry.is_true,
-                    publisher_get_last_node(context),
-                    current_key,
-                );
-                // END PUBLISHER
-                return (
-                    memo_entry.is_true,
-                    memo_entry.position,
-                    memo_entry.ast_or_lr.into(),
-                );
-            }
-            None => {}
-        }
-    }
-    // If not cached value we just run the func.
-
-    let (parent_key, current_key) = publisher_setup_node(context, rule);
-    println!("{:?} Entering No Cached Value Block", rule);
-    {
-        // Necessary modificaiton to deny LR.
-        let mut cache = context.cache.borrow_mut();
-        cache.push(rule, false, position, 0, cache::ASTOrLR::AST(AST::FAIL))
-    }
-    let ans = func(context, source, position);
-    publisher_update_node(context, position, ans.1, ans.0, parent_key, current_key);
-    let mut cache = context.cache.borrow_mut();
-    cache.push(
-        rule,
-        ans.0,
-        position,
-        ans.1,
-        cache::ASTOrLR::AST(AST::SUCCESS(current_key)),
-    );
-    println!("{:?} Exiting No Cached Value Block", rule);
-    println!("{:?} Exiting Var_Name Kernel", rule);
-    return ans;
-}
 
 fn setup_lr_var_name_kernel<T: Cache, S: Publisher>(
     rule: Rules,
@@ -260,7 +57,7 @@ fn setup_lr_var_name_kernel<T: Cache, S: Publisher>(
     }
 }
 
-fn _var_name_kernel_direct_lr<T: Cache, S: Publisher>(
+pub fn _var_name_kernel_indirect_lr<T: Cache, S: Publisher>(
     rule: Rules,
     context: &Context<T, S>,
     source: &Source,
